refactor: Line Domain 원시값 포장 및 메소드 분리
